+++ app/src/main/java/com/keylesspalace/tusky/components/compose/ComposeViewModel.kt
@@ -21,7 +21,6 @@ import androidx.core.net.toUri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModel
 import com.keylesspalace.tusky.adapter.ComposeAutoCompleteAdapter
 import com.keylesspalace.tusky.components.compose.ComposeActivity.QueuedMedia
 import com.keylesspalace.tusky.components.search.SearchType
@@ -34,22 +33,10 @@ import com.keylesspalace.tusky.network.MastodonApi
 import com.keylesspalace.tusky.service.ServiceClient
 import com.keylesspalace.tusky.service.TootToSend
 import com.keylesspalace.tusky.util.*
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.Singles
 import java.util.*
 import javax.inject.Inject
-
-open class RxAwareViewModel : ViewModel() {
-    private val disposables = CompositeDisposable()
-
-    fun Disposable.autoDispose() = disposables.add(this)
-
-    override fun onCleared() {
-        super.onCleared()
-        disposables.clear()
-    }
-}
 
 /**
  * Throw when trying to add an image when video is already present or the other way around
